refactor: Extract writer methods to FuseSessionExt trait
To reduce code duplication between the Linux and macOS FuseSession
implementations, this change extracts the with_writer and
try_with_writer helper methods into a new extension trait,
FuseSessionExt.

The trait provides default implementations for these methods, which rely
on file() and bufsize() accessor methods that the concrete session types
must provide. Both linux_session::FuseSession and
macos_session::FuseSession now implement this trait, allowing them to
share the common logic.
