+++ src/transport/fusedev/linux_session.rs
@@ -24,6 +24,8 @@ use nix::poll::{poll, PollFd, PollFlags};
 use nix::sys::epoll::{epoll_ctl, EpollEvent, EpollFlags, EpollOp};
 use nix::unistd::{getgid, getuid, read};
 
+use crate::transport::fusedev::FuseSessionExt;
+
 use super::{
     super::pagesize,
     Error::{IoError, SessionFailure},
@@ -238,35 +240,6 @@ impl FuseSession {
         }
     }
 
-    /// Create a new fuse message writer and pass it to the given closure.
-    pub fn with_writer<F>(&mut self, f: F)
-    where
-        F: FnOnce(FuseDevWriter),
-    {
-        if let Some(file) = &self.file {
-            let fd = file.as_raw_fd();
-            let mut buf = vec![0x0u8; self.bufsize];
-            let writer = FuseDevWriter::new(fd, &mut buf).unwrap();
-            f(writer);
-        }
-    }
-
-    /// Create a new fuse message writer and pass it to the given closure. and return the result from the closure.
-    pub fn try_with_writer<F, R, E>(&mut self, f: F) -> std::result::Result<R, E>
-    where
-        F: FnOnce(FuseDevWriter) -> std::result::Result<R, E>,
-        E: From<super::Error>,
-    {
-        if let Some(file) = &self.file {
-            let fd = file.as_raw_fd();
-            let mut buf = vec![0x0u8; self.bufsize];
-            let writer = FuseDevWriter::new(fd, &mut buf)?;
-            f(writer)
-        } else {
-            Err(SessionFailure("invalid fuse session".into()).into())
-        }
-    }
-
     /// Wake channel loop and exit
     pub fn wake(&self) -> Result<()> {
         let wakers = self
@@ -294,6 +267,16 @@ impl FuseSession {
 impl Drop for FuseSession {
     fn drop(&mut self) {
         let _ = self.umount();
+    }
+}
+
+impl FuseSessionExt for FuseSession {
+    fn file(&self) -> Option<&File> {
+        self.file.as_ref()
+    }
+
+    fn bufsize(&self) -> usize {
+        self.bufsize
     }
 }
